ctrldev: move UI coupled methods to a new module
