+++ zyngine/ctrldev/zynthian_ctrldev_akai_apc_key25_mk2.py
@@ -33,14 +33,16 @@ from bisect import bisect
 from functools import partial
 from copy import deepcopy
 from threading import Thread, RLock, Event
-from zyngine.ctrldev.zynthian_ctrldev_base import (
-    zynthian_ctrldev_zynmixer, zynthian_ctrldev_zynpad, RunTimer, ModeHandlerBase,
-    KnobSpeedControl, ButtonTimer, CONST
-)
 from zyngine.zynthian_engine_audioplayer import zynthian_engine_audioplayer
 from zyngine.zynthian_signal_manager import zynsigman
 from zyncoder.zyncore import lib_zyncore
 from zynlibs.zynseq import zynseq
+
+from .zynthian_ctrldev_base import (
+    zynthian_ctrldev_zynmixer, zynthian_ctrldev_zynpad, RunTimer,
+    KnobSpeedControl, ButtonTimer, CONST
+)
+from .zynthian_ctrldev_base_ui import ModeHandlerBase
 
 
 # FIXME: these defines should be taken from where they are defined (zynseq.h)
